importfile double conversion

+++ importfile.m
@@ -9,6 +9,6 @@ newData1 = load('-mat', fileToRead);
 % Create new variables in the base workspace from those fields
 vars = fieldnames(newData1);
 for i = 1:length(vars)
-    assignin('base', vars{i}, newData1.(vars{i}));
+    assignin('base', vars{i}, double(newData1.(vars{i})));
 end
